refactor(rust): extract inline tests to tests/ directory and sync with JS coverage
- Create lib.rs to expose modules for integration testing
- Extract all #[cfg(test)] inline tests from 21 src files into 22 test files in tests/
- Add new test cases to match JavaScript integration test coverage
- Sync bash tool metadata with JS format (add output/exit fields)
- Add Prefix::Call variant to id module for JS compatibility
- Make private helper functions pub for external test access:
  - edit.rs: all 9 replacement strategy functions
  - grep.rs: matches_glob
  - list.rs: should_ignore
  - webfetch.rs: extract_text_from_html, decode_html_entities, strip_remaining_tags
  - binary.rs: is_binary_extension
- Fix all clippy warnings (unused imports, idiomatic patterns)
- All 236 tests pass, clippy clean, fmt clean
