+++ rust/Cargo.toml
@@ -67,6 +67,10 @@ tempfile = "3.14"
 assert_cmd = "2.0"
 predicates = "3.1"
 
+[lib]
+name = "agent"
+path = "src/lib.rs"
+
 [[bin]]
 name = "agent"
 path = "src/main.rs"
