+++ rust/src/error.rs
@@ -174,38 +174,3 @@ impl AgentError {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_file_not_found_error() {
-        let err = AgentError::file_not_found(
-            "/path/to/file.txt",
-            vec![
-                "/path/to/file.ts".to_string(),
-                "/path/to/file.js".to_string(),
-            ],
-        );
-
-        let json = err.to_json();
-        assert_eq!(json["name"], "FileNotFound");
-        assert_eq!(json["data"]["path"], "/path/to/file.txt");
-    }
-
-    #[test]
-    fn test_invalid_arguments_error() {
-        let err = AgentError::invalid_arguments("read", "filePath is required");
-
-        let json = err.to_json();
-        assert_eq!(json["name"], "InvalidArguments");
-        assert_eq!(json["data"]["tool"], "read");
-    }
-
-    #[test]
-    fn test_error_display() {
-        let err = AgentError::file_not_found("/test.txt", vec![]);
-        assert_eq!(format!("{err}"), "File not found: /test.txt");
-    }
-}
