+++ rust/src/tool/batch.rs
@@ -238,85 +238,3 @@ impl Tool for BatchTool {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use tempfile::TempDir;
-
-    fn create_context(dir: &std::path::Path) -> ToolContext {
-        ToolContext::new("ses_test", "msg_test", dir)
-    }
-
-    #[tokio::test]
-    async fn test_batch_disallowed_tool() {
-        let temp = TempDir::new().unwrap();
-        let tool = BatchTool;
-        let ctx = create_context(temp.path());
-
-        let params = json!({
-            "tool_calls": [
-                {
-                    "tool": "batch",
-                    "parameters": {}
-                }
-            ]
-        });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-        assert!(result.metadata["failed"].as_u64().unwrap() > 0);
-    }
-
-    #[tokio::test]
-    async fn test_batch_unknown_tool() {
-        let temp = TempDir::new().unwrap();
-        let tool = BatchTool;
-        let ctx = create_context(temp.path());
-
-        let params = json!({
-            "tool_calls": [
-                {
-                    "tool": "nonexistent_tool",
-                    "parameters": {}
-                }
-            ]
-        });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-        assert!(result.metadata["failed"].as_u64().unwrap() > 0);
-    }
-
-    #[tokio::test]
-    async fn test_batch_multiple_reads() {
-        let temp = TempDir::new().unwrap();
-        let file1 = temp.path().join("file1.txt");
-        let file2 = temp.path().join("file2.txt");
-        fs::write(&file1, "content 1").unwrap();
-        fs::write(&file2, "content 2").unwrap();
-
-        let tool = BatchTool;
-        let ctx = create_context(temp.path());
-
-        let params = json!({
-            "tool_calls": [
-                {
-                    "tool": "read",
-                    "parameters": {
-                        "filePath": file1.to_string_lossy()
-                    }
-                },
-                {
-                    "tool": "read",
-                    "parameters": {
-                        "filePath": file2.to_string_lossy()
-                    }
-                }
-            ]
-        });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-        assert_eq!(result.metadata["successful"].as_u64().unwrap(), 2);
-        assert_eq!(result.metadata["failed"].as_u64().unwrap(), 0);
-    }
-}
