+++ rust/src/tool/multiedit.rs
@@ -121,45 +121,3 @@ impl Tool for MultiEditTool {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use tempfile::TempDir;
-
-    fn create_context(dir: &std::path::Path) -> ToolContext {
-        ToolContext::new("ses_test", "msg_test", dir)
-    }
-
-    #[tokio::test]
-    async fn test_multiedit_sequential() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("test.txt");
-        fs::write(&file_path, "hello world foo bar").unwrap();
-
-        let tool = MultiEditTool;
-        let ctx = create_context(temp.path());
-
-        let params = json!({
-            "filePath": file_path.to_string_lossy(),
-            "edits": [
-                {
-                    "filePath": file_path.to_string_lossy(),
-                    "oldString": "hello",
-                    "newString": "greetings"
-                },
-                {
-                    "filePath": file_path.to_string_lossy(),
-                    "oldString": "foo",
-                    "newString": "baz"
-                }
-            ]
-        });
-
-        tool.execute(params, &ctx).await.unwrap();
-
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert_eq!(content, "greetings world baz bar");
-    }
-}
