+++ rust/src/tool/read.rs
@@ -261,78 +261,3 @@ async fn read_image(
         attachments: Some(vec![attachment]),
     })
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn create_context(dir: &Path) -> ToolContext {
-        ToolContext::new("ses_test", "msg_test", dir)
-    }
-
-    #[tokio::test]
-    async fn test_read_text_file() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("test.txt");
-        fs::write(&file_path, "line 1\nline 2\nline 3\n").unwrap();
-
-        let tool = ReadTool;
-        let ctx = create_context(temp.path());
-        let params = json!({ "filePath": file_path.to_string_lossy() });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-
-        assert!(result.output.contains("line 1"));
-        assert!(result.output.contains("line 2"));
-        assert!(result.output.contains("line 3"));
-        assert!(result.output.contains("00001|"));
-    }
-
-    #[tokio::test]
-    async fn test_read_with_offset_and_limit() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("test.txt");
-        let content: String = (1..=100).map(|i| format!("line {}\n", i)).collect();
-        fs::write(&file_path, content).unwrap();
-
-        let tool = ReadTool;
-        let ctx = create_context(temp.path());
-        let params = json!({
-            "filePath": file_path.to_string_lossy(),
-            "offset": 10,
-            "limit": 5
-        });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-
-        assert!(result.output.contains("line 11"));
-        assert!(result.output.contains("line 15"));
-        assert!(!result.output.contains("line 16"));
-    }
-
-    #[tokio::test]
-    async fn test_read_nonexistent_file() {
-        let temp = TempDir::new().unwrap();
-        let tool = ReadTool;
-        let ctx = create_context(temp.path());
-        let params = json!({ "filePath": "/nonexistent/file.txt" });
-
-        let result = tool.execute(params, &ctx).await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_read_binary_file() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("test.bin");
-        fs::write(&file_path, &[0, 1, 2, 3, 0, 0, 0]).unwrap();
-
-        let tool = ReadTool;
-        let ctx = create_context(temp.path());
-        let params = json!({ "filePath": file_path.to_string_lossy() });
-
-        let result = tool.execute(params, &ctx).await;
-        assert!(result.is_err());
-    }
-}
