+++ rust/src/tool/todo.rs
@@ -173,69 +173,3 @@ impl Tool for TodoReadTool {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn create_context(dir: &std::path::Path, session_id: &str) -> ToolContext {
-        ToolContext::new(session_id, "msg_test", dir)
-    }
-
-    #[tokio::test]
-    async fn test_todo_write_and_read() {
-        let temp = TempDir::new().unwrap();
-        let ctx = create_context(temp.path(), "ses_test_todo_write_read");
-
-        let write_tool = TodoWriteTool;
-        let params = json!({
-            "todos": [
-                {
-                    "content": "Fix bug",
-                    "status": "pending",
-                    "activeForm": "Fixing bug"
-                },
-                {
-                    "content": "Write tests",
-                    "status": "in_progress",
-                    "activeForm": "Writing tests"
-                }
-            ]
-        });
-
-        let write_result = write_tool.execute(params, &ctx).await.unwrap();
-        assert!(write_result.title.contains("2 todos"));
-
-        let read_tool = TodoReadTool;
-        let read_result = read_tool.execute(json!({}), &ctx).await.unwrap();
-        assert!(read_result.output.contains("Fix bug"));
-        assert!(read_result.output.contains("Write tests"));
-    }
-
-    #[tokio::test]
-    async fn test_todo_completed_count() {
-        let temp = TempDir::new().unwrap();
-        let ctx = create_context(temp.path(), "ses_test_todo_completed_count");
-
-        let write_tool = TodoWriteTool;
-        let params = json!({
-            "todos": [
-                {
-                    "content": "Done task",
-                    "status": "completed",
-                    "activeForm": "Completing task"
-                },
-                {
-                    "content": "Pending task",
-                    "status": "pending",
-                    "activeForm": "Doing pending task"
-                }
-            ]
-        });
-
-        let result = write_tool.execute(params, &ctx).await.unwrap();
-        // Only 1 non-completed todo
-        assert!(result.title.contains("1 todos"));
-    }
-}
