+++ rust/src/tool/websearch.rs
@@ -84,6 +84,7 @@ struct McpSearchResult {
 #[derive(Debug, Deserialize)]
 struct McpContent {
     #[serde(rename = "type")]
+    #[allow(dead_code)]
     content_type: String,
     text: String,
 }
@@ -208,22 +209,5 @@ impl Tool for WebSearchTool {
             metadata: json!({}),
             attachments: None,
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn create_context(dir: &std::path::Path) -> ToolContext {
-        ToolContext::new("ses_test", "msg_test", dir)
-    }
-
-    #[test]
-    fn test_params_schema() {
-        let tool = WebSearchTool;
-        let schema = tool.parameters_schema();
-        assert!(schema["properties"]["query"].is_object());
     }
 }
