+++ rust/src/tool/write.rs
@@ -91,73 +91,3 @@ impl Tool for WriteTool {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs as std_fs;
-    use tempfile::TempDir;
-
-    fn create_context(dir: &std::path::Path) -> ToolContext {
-        ToolContext::new("ses_test", "msg_test", dir)
-    }
-
-    #[tokio::test]
-    async fn test_write_new_file() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("new_file.txt");
-
-        let tool = WriteTool;
-        let ctx = create_context(temp.path());
-        let params = json!({
-            "content": "Hello, World!",
-            "filePath": file_path.to_string_lossy()
-        });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-
-        assert!(file_path.exists());
-        assert_eq!(std_fs::read_to_string(&file_path).unwrap(), "Hello, World!");
-        assert_eq!(result.metadata["exists"], false);
-    }
-
-    #[tokio::test]
-    async fn test_write_overwrite_file() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("existing.txt");
-        std_fs::write(&file_path, "old content").unwrap();
-
-        let tool = WriteTool;
-        let ctx = create_context(temp.path());
-        let params = json!({
-            "content": "new content",
-            "filePath": file_path.to_string_lossy()
-        });
-
-        let result = tool.execute(params, &ctx).await.unwrap();
-
-        assert_eq!(std_fs::read_to_string(&file_path).unwrap(), "new content");
-        assert_eq!(result.metadata["exists"], true);
-    }
-
-    #[tokio::test]
-    async fn test_write_creates_directories() {
-        let temp = TempDir::new().unwrap();
-        let file_path = temp.path().join("a").join("b").join("c").join("file.txt");
-
-        let tool = WriteTool;
-        let ctx = create_context(temp.path());
-        let params = json!({
-            "content": "nested content",
-            "filePath": file_path.to_string_lossy()
-        });
-
-        tool.execute(params, &ctx).await.unwrap();
-
-        assert!(file_path.exists());
-        assert_eq!(
-            std_fs::read_to_string(&file_path).unwrap(),
-            "nested content"
-        );
-    }
-}
